requirements.txt
numpy
pyglet
pyinstaller
pypresence
pillow
colorama